requirements.txt
langchain==1.2.15
langchain-community==0.4.1
langchain-core==1.2.28
langchain-text-splitters==1.1.1
langgraph==1.1.6
chromadb
dashscope
python-dotenv
pytest
streamlit

# 监控和评估依赖
langsmith>=0.1.0  # LangSmith SDK for tracing and monitoring
structlog>=24.1.0  # Structured logging
prometheus-client>=0.20.0  # Metrics collection (optional)
psutil>=5.9.0  # System metrics collection
python-json-logger>=3.0.0  # JSON logging support
rich>=13.7.0  # Console output formatting
watchdog>=3.0.0  # File system monitoring (optional)